+++ install/scripts/phase2/080karaf.sh
@@ -7,6 +7,12 @@ if [ -f "$HOME_DIR/islandora/install/configs/variables" ]; then
   . "$HOME_DIR"/islandora/install/configs/variables
 fi
 
+
+set -e
+if [ -f ~/karaf ]; then
+    exit
+fi
+
 if [ ! -f "$DOWNLOAD_DIR/apache-karaf-$KARAF_VERSION.tar.gz" ]; then
   echo "Downloading Karaf"
   wget -q -O "$DOWNLOAD_DIR/apache-karaf-$KARAF_VERSION.tar.gz" "http://mirror.csclub.uwaterloo.ca/apache/karaf/$KARAF_VERSION/apache-karaf-$KARAF_VERSION.tar.gz"
@@ -15,9 +21,15 @@ fi
 cd "$HOME_DIR"
 
 echo "Download and install Karaf"
-cp -v "$DOWNLOAD_DIR/apache-karaf-$KARAF_VERSION.tar.gz" "$HOME_DIR"
+if [ -f "/opt/karaf/bin/stop" ]; then
+    /opt/karaf/bin/stop
+    sleep 10
+    rm /opt/karaf
+fi
+cp  "$DOWNLOAD_DIR/apache-karaf-$KARAF_VERSION.tar.gz" "$HOME_DIR"
 tar zxf apache-karaf-"$KARAF_VERSION".tar.gz
 rm apache-karaf-"$KARAF_VERSION".tar.gz
+rm -rf /opt/apache-karaf-"$KARAF_VERSION"
 mv apache-karaf-"$KARAF_VERSION" /opt
 ln -s /opt/apache-karaf-"$KARAF_VERSION" /opt/karaf
 
@@ -28,6 +40,9 @@ sleep 60
 /opt/karaf/bin/stop
 
 echo "Add Karaf as a Linux service"
+if [ -f "/etc/init.d/karaf-service" ]; then
+  rm "/etc/init.d/karaf-service"
+fi
 ln -s /opt/karaf/bin/karaf-service /etc/init.d/
 update-rc.d karaf-service defaults
 
@@ -40,3 +55,5 @@ sed -i "s|#org.ops4j.pax.url.mvn.localRepository=|org.ops4j.pax.url.mvn.localRep
 echo "Start Karaf"
 service karaf-service start
 sleep 60
+
+touch ~/karaf
